+++ src/startup.cpp
@@ -113,9 +113,9 @@ Menu newgameMenu {
 
 // https://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function
 unsigned long hashString(const std::string &str) {
-    const unsigned long FNV_offset_basis = 14695981039346656037ul;
-    const unsigned long FNV_prime = 1099511628211ul;
-    unsigned long hash = FNV_offset_basis;
+    const unsigned long long FNV_offset_basis = 14695981039346656037ul;
+    const unsigned long long FNV_prime = 1099511628211ul;
+    unsigned long long hash = FNV_offset_basis;
 
     for (char c : str) {
         hash ^= static_cast<unsigned long>(c);
